Majority of code done.
Renamed the files to their proper names. Linked in the css files. Linked in the javascript files. Displayed the table to the web page. Displayed the list of words. Added in a way to select words for the word search. Added in the code to the solution list. Added in the code to check if someone one. Added in the code to check if someone has found a word. Added in the code to cross of the words when found on the list. Added in the majority of comments to the code.